Refactor to prepare for rest api
